web/requirements.txt
pandas~=1.2.2
Flask~=1.1.2
PyMySQL~=1.0.2
toolforge~=4.4.0
PyYAML~=5.4.1
numpy~=1.20.1